+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support for RubyMine 2018.01 (via @kibitan)
 - Support for IntelliJ Idea 2017.1, 2017.2, 2017.3, and 2018.1 (via @cool00geek)
 - Remove references to OpenSSH private key syncing - removed in dcb26ba (via @njdancer)
+- Add logic to detect mackup config in iCloud (via @njdancer)
 
 ## Mackup 0.8.18